lots of player folder refactoring in the name of making this area of the codebase more workable
