+++ src/components/CarPlay/Album.tsx
@@ -1,7 +1,7 @@
 import { BaseItemDto } from '@jellyfin/sdk/lib/generated-client'
 import { CarPlay, ListTemplate } from 'react-native-carplay'
 import CarPlayNowPlaying from './NowPlaying'
-import { loadNewQueue } from '../../hooks/player/functions/queue'
+import { loadNewQueue } from '../../player/queuing'
 import { formatArtistNames } from '../../utils/formatting/artist-names'
 
 const AlbumTemplate = (
