+++ src/components/CarPlay/Tracks.tsx
@@ -4,7 +4,7 @@ import uuid from 'react-native-uuid'
 import CarPlayNowPlaying from './NowPlaying'
 import { Queue } from '../../services/types/queue-item'
 import AlbumTemplate from './Album'
-import { loadNewQueue } from '../../hooks/player/functions/queue'
+import { loadNewQueue } from '../../player/queuing'
 import { ensureAlbumDiscsQuery } from '../../api/queries/album'
 import { formatArtistNames } from '../../utils/formatting/artist-names'
 import { getItemImageUrl } from '../../api/queries/image/utils'
