+++ src/components/Context/index.tsx
@@ -11,7 +11,7 @@ import { useQuery } from '@tanstack/react-query'
 import { QueryKeys } from '../../enums/query-keys'
 import { fetchAlbumDiscs } from '../../api/queries/item'
 import { getItemsApi } from '@jellyfin/sdk/lib/utils/api'
-import { AddToQueueMutation } from '../../hooks/player/interfaces'
+import { AddToQueueMutation } from '../../player/interfaces'
 import { QueuingType } from '../../enums/queuing-type'
 import { useEffect } from 'react'
 import navigationRef from '../../screens/navigation'
@@ -28,7 +28,7 @@ import { useAreAllDownloaded } from '../../hooks/downloads'
 import { useDownloadProgress } from 'react-native-nitro-player'
 import CircularProgressIndicator from '../Global/components/circular-progress-indicator'
 import { useArtist } from '../../api/queries/artist'
-import { addToQueue } from '../../hooks/player/functions/queue'
+import { addToQueue } from '../../player/queuing'
 import { useAlbum } from '../../api/queries/album'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import ViewInstantMixMenuRow from './components/instant-mix-row'
