+++ src/components/Library/tab-bar.tsx
@@ -4,7 +4,7 @@ import { XStack, YStack, Paragraph } from 'tamagui'
 import Icon from '../Global/components/icon'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import useLibraryStore from '../../stores/library'
-import { handleLibraryShuffle } from '../../hooks/player/functions/shuffle'
+import { handleLibraryShuffle } from '../../player/controls/shuffle'
 import { TrackPlayer } from 'react-native-nitro-player'
 import { useNavigation } from '@react-navigation/native'
 import { NativeStackNavigationProp } from '@react-navigation/native-stack'
