--- src/hooks/player/functions/queue.ts
@@ -1,325 +1,0 @@
-import { mapDtosToTracks } from '../../../utils/mapping/item-to-track'
-import { networkStatusTypes } from '../../../components/Network/internetConnectionWatcher'
-import { clearPlaylists, filterTracksOnNetworkStatus } from './utils/queue'
-import { AddToQueueMutation, QueueMutation, QueueOrderMutation } from '../interfaces'
-import { shuffleJellifyTracks } from './utils/shuffle'
-import { setNewQueue, usePlayerQueueStore } from '../../../stores/player/queue'
-import { isNull } from 'lodash'
-import { useNetworkStore } from '../../../stores/network'
-import { PlayerQueue, TrackItem, TrackPlayer } from 'react-native-nitro-player'
-import uuid from 'react-native-uuid'
-import Toast from 'react-native-toast-message'
-import { QueuingType } from '../../../enums/queuing-type'
-import { ensureDownloadedTracks } from '../../downloads/utils'
-import { applyHapticFeedback } from '../../../utils/haptics'
-import { setPlaybackPosition } from '../../../stores/player/playback'
-
-type LoadQueueResult = {
-	finalStartIndex: number
-	tracks: TrackItem[]
-}
-
-export const loadNewQueue = async (variables: QueueMutation) => {
-	applyHapticFeedback('info')
-
-	await loadQueue({ ...variables })
-
-	if (variables.startPlayback) {
-		await TrackPlayer.play()
-	}
-}
-
-async function loadQueue({
-	index = 0,
-	tracklist,
-	queue,
-	shuffled = false,
-}: QueueMutation): Promise<LoadQueueResult> {
-	await TrackPlayer.pause()
-
-	const networkStatus = useNetworkStore.getState().networkStatus ?? networkStatusTypes.ONLINE
-
-	// Get the item at the start index
-	const startingTrack = tracklist[index]
-
-	const downloadedTracks = await ensureDownloadedTracks()
-
-	const availableAudioItems = filterTracksOnNetworkStatus(
-		networkStatus as networkStatusTypes,
-		tracklist,
-		downloadedTracks ?? [],
-	)
-
-	// Convert to JellifyTracks first
-	let playlist = mapDtosToTracks(availableAudioItems, downloadedTracks)
-
-	// Store the original unshuffled queue
-	usePlayerQueueStore.getState().setUnshuffledQueue(playlist)
-
-	// Handle if a shuffle was requested
-	if (shuffled && playlist.length > 1) {
-		const startingTrackId = startingTrack.Id
-		const mappedStartingTrack = playlist.find((track) => track.id === startingTrackId)
-
-		if (mappedStartingTrack) {
-			const remainingTracks = playlist.filter((track) => track.id !== startingTrackId)
-			const { shuffled: shuffledTracks } = shuffleJellifyTracks(remainingTracks)
-			playlist = [mappedStartingTrack, ...shuffledTracks]
-		} else {
-			const { shuffled: shuffledTracks } = shuffleJellifyTracks(playlist)
-			playlist = shuffledTracks
-		}
-	}
-
-	const rawStartIndex = playlist.findIndex((item) => item.id === startingTrack.Id)
-	const finalStartIndex = rawStartIndex >= 0 ? rawStartIndex : 0
-
-	await clearPlaylists()
-
-	const playlistId = await PlayerQueue.createPlaylist(uuid.v4(), undefined, undefined)
-
-	await PlayerQueue.addTracksToPlaylist(playlistId, playlist)
-
-	setNewQueue(playlist, queue, finalStartIndex, shuffled)
-	setPlaybackPosition(0)
-
-	await PlayerQueue.loadPlaylist(playlistId, finalStartIndex)
-
-	return {
-		finalStartIndex,
-		tracks: playlist,
-	}
-}
-
-/**
- * Inserts a track at the next index in the queue
- *
- * Keeps a copy of the original queue in {@link unshuffledQueue}
- *
- * @param item The track to play next
- */
-export const playNextInQueue = async ({ tracks }: AddToQueueMutation) => {
-	const { currentIndex, queue } = usePlayerQueueStore.getState()
-
-	const queuedIds = queue.map((track) => track.id)
-
-	const insertIndex = calculatePlayNextInsertIndex(currentIndex, queue)
-
-	const downloadedTracks = await ensureDownloadedTracks()
-
-	const newTracks = mapDtosToTracks(tracks, downloadedTracks)
-
-	const playlistId = await PlayerQueue.getCurrentPlaylistId()
-
-	if (isNull(playlistId)) {
-		console.warn('playNextInQueue: No active playlist to add to')
-		return
-	}
-
-	let updatedIndex = currentIndex
-	const tracksToReorder: TrackItem[] = []
-	const tracksToAdd: TrackItem[] = []
-
-	newTracks.forEach((track) => {
-		if (queuedIds.includes(track.id)) {
-			tracksToReorder.push(track)
-		} else {
-			tracksToAdd.push(track)
-		}
-	})
-
-	// Reorder existing tracks to the next position
-	if (tracksToReorder.length > 0) {
-		const tracksMovingPastCurrent = tracksToReorder.filter((track) => {
-			const currentPos = queue.findIndex((t) => t.id === track.id)
-			return currentPos < (currentIndex ?? 0)
-		}).length
-
-		const reorderPromises = tracksToReorder.map((track, index) =>
-			PlayerQueue.reorderTrackInPlaylist(
-				playlistId,
-				track.id,
-				insertIndex + index - tracksMovingPastCurrent,
-			),
-		)
-		await Promise.all(reorderPromises)
-
-		// Update the current index if it's changed
-		updatedIndex = await TrackPlayer.getCurrentTrackIndex()
-	}
-
-	// Add new tracks to the queue
-	if (tracksToAdd.length > 0) {
-		await PlayerQueue.addTracksToPlaylist(playlistId, tracksToAdd, insertIndex)
-	}
-
-	const updatedQueue = await TrackPlayer.getActualQueue()
-	usePlayerQueueStore.setState((state) => ({
-		...state,
-		currentIndex: updatedIndex,
-		queue: [...updatedQueue],
-		unShuffledQueue: [...state.unShuffledQueue, ...newTracks],
-	}))
-}
-
-export const playLaterInQueue = async ({ tracks }: AddToQueueMutation) => {
-	const downloadedTracks = await ensureDownloadedTracks()
-
-	const newTracks = mapDtosToTracks(tracks, downloadedTracks)
-
-	const playlistId = await PlayerQueue.getCurrentPlaylistId()
-
-	if (isNull(playlistId)) {
-		console.warn('playLaterInQueue: No active playlist to add to')
-		return
-	}
-
-	// Add to the end of the queue
-	await PlayerQueue.addTracksToPlaylist(playlistId, newTracks)
-
-	// Get the active queue and update Zustand while isQueuing=true blocks callbacks
-	const updatedQueue = await TrackPlayer.getActualQueue()
-
-	usePlayerQueueStore.setState((state) => ({
-		...state,
-		queue: updatedQueue,
-		unShuffledQueue: [...state.unShuffledQueue, ...newTracks],
-	}))
-}
-
-export const addToQueue = async (variables: AddToQueueMutation) => {
-	try {
-		const actualQueue = await TrackPlayer.getActualQueue()
-		const actualQueueIds = actualQueue.map((t) => t.id)
-
-		if (variables.queuingType === QueuingType.PlayNext) {
-			// For PlayNext, pass all tracks so we can reorder existing ones
-			await playNextInQueue(variables)
-		} else {
-			// For PlayLater, only add new tracks
-			const tracksToAdd = variables.tracks.filter(
-				(item) => !actualQueueIds.includes(item.Id!),
-			)
-			await playLaterInQueue({ ...variables, tracks: tracksToAdd })
-		}
-
-		applyHapticFeedback('success')
-		Toast.show({
-			text1:
-				variables.queuingType === QueuingType.PlayNext ? 'Playing next' : 'Added to queue',
-			type: 'success',
-		})
-	} catch (error) {
-		applyHapticFeedback('error')
-		console.error(
-			`Failed to ${variables.queuingType === QueuingType.PlayNext ? 'play next' : 'add to queue'}`,
-			error,
-		)
-		Toast.show({
-			text1:
-				variables.queuingType === QueuingType.PlayNext
-					? 'Failed to play next'
-					: 'Failed to add to queue',
-			type: 'error',
-		})
-	}
-}
-
-export const removeItemFromQueue = async (index: number) => {
-	applyHapticFeedback('info')
-
-	const playlistId = PlayerQueue.getCurrentPlaylistId()
-
-	if (!playlistId) return
-
-	const playlist = PlayerQueue.getPlaylist(playlistId)!
-	const trackIdToRemove = playlist.tracks[index].id
-
-	// Remove from the native playlist first. If this is the currently-playing track,
-	// the native rebuildQueueFromCurrentPosition will detect the mismatch and
-	// automatically advance playback to the next track via playFromIndexInternal /
-	// rebuildQueueFromPlaylistIndex — no explicit skipToNext() needed.
-	PlayerQueue.removeTrackFromPlaylist(playlistId, trackIdToRemove)
-
-	const {
-		queue: prevQueue,
-		unShuffledQueue: prevUnshuffledQueue,
-		currentIndex,
-	} = usePlayerQueueStore.getState()
-
-	const newQueue = prevQueue.filter((_, i) => i !== index)
-
-	// Also remove from unShuffledQueue to prevent orphaned tracks
-	const newUnshuffledQueue = prevUnshuffledQueue.filter((t) => t.id !== trackIdToRemove)
-
-	// If queue is now empty, stop playback and tear down
-	if (newQueue.length === 0) {
-		TrackPlayer.pause()
-		usePlayerQueueStore.setState((state) => ({
-			...state,
-			queue: newQueue,
-			unShuffledQueue: newUnshuffledQueue,
-		}))
-		usePlayerQueueStore.getState().setCurrentIndex(undefined)
-		PlayerQueue.deletePlaylist(playlistId)
-		return
-	}
-
-	// If a track before the current one was removed, shift the index down so it
-	// keeps pointing at the same still-playing track.
-	const newCurrentIndex = index < (currentIndex ?? 0) ? (currentIndex ?? 0) - 1 : currentIndex
-
-	usePlayerQueueStore.setState((state) => ({
-		...state,
-		queue: newQueue,
-		unShuffledQueue: newUnshuffledQueue,
-		currentIndex: newCurrentIndex,
-	}))
-}
-
-export const reorderQueue = async ({ fromIndex, toIndex }: QueueOrderMutation) => {
-	const playlistId = PlayerQueue.getCurrentPlaylistId()
-
-	if (!playlistId) return
-
-	const { queue: prevQueue, currentIndex: prevIndex } = usePlayerQueueStore.getState()
-	const currentTrack = prevIndex !== undefined ? prevQueue[prevIndex] : undefined
-
-	const { tracks } = PlayerQueue.getPlaylist(playlistId)!
-
-	await PlayerQueue.reorderTrackInPlaylist(playlistId, tracks[fromIndex].id, toIndex)
-
-	const queue = await TrackPlayer.getActualQueue()
-	const updatedCurrentIndex = currentTrack
-		? queue.findIndex((track) => track.id === currentTrack.id)
-		: -1
-
-	usePlayerQueueStore.setState((state) => ({
-		...state,
-		queue,
-		currentIndex: updatedCurrentIndex !== -1 ? updatedCurrentIndex : prevIndex,
-	}))
-}
-
-/**
- * Calculates the insert index for the new tracks.
- *
- * If there is a current track and the queue has at least one track after the current track, insert after the current track.
- *
- * If there is a current track at the end of the queue, insert at the end of the queue.
- *
- * If there is no current track, insert at the start of the queue.
- */
-function calculatePlayNextInsertIndex(currentIndex: number | undefined, queue: TrackItem[]) {
-	let insertIndex = 0
-
-	if (currentIndex === undefined) return insertIndex
-
-	if (currentIndex < queue.length - 1) {
-		insertIndex = currentIndex + 1
-	} else {
-		insertIndex = queue.length
-	}
-
-	return insertIndex
-}
