+++ src/player/queuing/types.ts
@@ -1,0 +1,6 @@
+import { TrackItem } from 'react-native-nitro-player'
+
+export type LoadQueueResult = {
+	finalStartIndex: number
+	tracks: TrackItem[]
+}
