+++ src/player/utils/queue.ts
@@ -1,6 +1,6 @@
 import { isNull, isUndefined } from 'lodash'
 import { BaseItemDto } from '@jellyfin/sdk/lib/generated-client/models'
-import { networkStatusTypes } from '../../../../components/Network/internetConnectionWatcher'
+import { networkStatusTypes } from '../../components/Network/internetConnectionWatcher'
 import { DownloadedTrack, PlayerQueue } from 'react-native-nitro-player'
 
 export async function clearPlaylists() {
